+++ game/public.h
@@ -57,7 +57,7 @@ typedef struct gameImportFuncs_s {
 typedef struct gameExportFuncs_s {
 	void(*Init)(void *clientInfo, void *imGuiContext);
 	void(*Console)(const char *line);
-	void(*Frame)(float dt);
+	void(*Frame)(double dt);
 } gameExportFuncs_t;
 
 extern gameImportFuncs_t *trap;
